test(logic): split 3 oversized resolver tests at group boundaries
Split three logic test files that exceeded the 400-line gate by
extracting coherent subgroups into sibling files. No assertions,
fixtures, or coverage are altered; each new file holds a strict
subset of the original groups.

- connectivity_resolver_test.dart: extracted sea/port scenarios into
  connectivity_resolver_sea_test.dart; original is now 'road/town'.
- naval_combat_resolver_test.dart: extracted resolveSeaBattle,
  applyNavalBattleResults, and navalInterceptProbability groups into
  naval_combat_resolver_resolution_test.dart.
- naval_test_part2_test.dart: extracted isWarpZoneSeaZone,
  region-scoped provinceIdsAdjacentToSeaZone, and fleetsInPortAtProvince
  groups into naval_test_part2_part2_test.dart, with a local topology
  setUp mirroring the original.

Drops 3 violations from the logic test file-size gate
(49 -> 46 over 400 lines on dev).

Refs #2216
